// admin/app/dashboard/foods/_components/types.ts
export interface Category {
  id: number;
  name: string;
  createdAt: string;
  updatedAt: string;
  foods: Food[];
}

export interface Food {
  id: number;
  name: string;
  price: string;
  image: string;
  ingredients: string;
  foodCategoyId: number;
  createdAt: string;
  updatedAt: string;
  category: Category;
}

export interface GetCategoriesResponse {
  categories: Category[];
}